design: observer for ProfileScreen
